add: package rename with app name
