+++ app/src/androidTest/java/io/github/network/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package io.github.design_system
+package io.github.network
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -19,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("io.github.design_system.test", appContext.packageName)
+        assertEquals("io.github.dawn", appContext.packageName)
     }
 }
