+++ app/src/main/AndroidManifest.xml
@@ -9,13 +9,13 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Template"
+        android:theme="@style/Theme.Dawn"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="io.github.network.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.Template">
+            android:theme="@style/Theme.Dawn">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
