+++ app/src/main/java/io/github/network/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package io.github.template.ui.theme
+package io.github.network.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -38,7 +38,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun TemplateTheme(
+fun DawnTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
