+++ app/src/main/java/io/github/network/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package io.github.template.ui.theme
+package io.github.network.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
