+++ app/src/test/java/io/github/network/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package io.github.design_system
+package io.github.network
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
