+++ common/design-system/src/androidTest/java/io/github/dawn/design_system/ExampleInstrumentedTest.kt
@@ -1,10 +1,12 @@
-package io.github.template
+package io.github.dawn.design_system
 
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert.assertEquals
+import androidx.test.ext.junit.runners.AndroidJUnit4
+
 import org.junit.Test
 import org.junit.runner.RunWith
+
+import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -17,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("io.github.template", appContext.packageName)
+        assertEquals("io.github.design_system.test", appContext.packageName)
     }
 }
