+++ common/network/src/main/java/io/github/dawn/network/MyClass.kt
@@ -1,0 +1,4 @@
+package io.github.dawn.network
+
+class MyClass {
+}
