--- common/network/src/main/java/io/github/network/MyClass.kt
@@ -1,4 +1,0 @@
-package io.github.network
-
-class MyClass {
-}
