+++ settings.gradle
@@ -12,7 +12,7 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "template"
+rootProject.name = "dawn"
 include ':app'
 include ':common:network'
 include ':common:design-system'
